+++ src/Functors/MoveParticles.h
@@ -23,8 +23,8 @@ class MoveParticles {
   const Coord3D boxMin;
   const int numCells[3];
   const int numCellsTotal;
-  const double cutoff;
-  const double deltaT;
+  const float cutoff;
+  const float deltaT;
   const BoundaryCondition boundaryCondition;
 
  public:
